Standardized all pixel index counters to type uint32_t in Transforms.cc and RawTile.h with added type-casting when necessary. Fixes CodeQL warnings about potential integer overflow.

+++ src/RawTile.h
@@ -79,7 +79,7 @@ class RawTile{
   int memoryManaged;
 
   /// The size of the data pointed to by data
-  size_t dataLength;
+  uint32_t dataLength;
 
   /// The width in pixels of this tile
   unsigned int width;
